track whether we had turned on climate control
The Leaf API doesn't give us the ability to query the current climate
state, but we do know that if we were asked to turn it on, it'll stay on
for about 15 minutes.  So track that state and reset it back to off 15
minutes later.  This lets us turn the climate control back off, and is a
decent estimation.